logic :Adapt the code according to the platform
