+++ cpp/fury/util/string_util_test.cc
@@ -17,54 +17,59 @@
  * under the License.
  */
 
-#include "string_util.h"
 #include <chrono>
-#include <gtest/gtest.h>
 #include <iostream>
+#include <random>
+
+#include "fury/util/logging.h"
+#include "string_util.h"
+#include "gtest/gtest.h"
 
 namespace fury {
 
-#if defined(__x86_64__) || defined(_M_X64)
-bool isLatin(const std::string &str) { return isLatin_AVX2(str); }
+// Function to generate a random string
+std::string generateRandomString(size_t length) {
+  const char charset[] =
+      "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
+  std::default_random_engine rng(std::random_device{}());
+  std::uniform_int_distribution<> dist(0, sizeof(charset) - 2);
 
-bool isLatin2(const std::string &str) { return isLatin_SSE2(str); }
-#elif defined(__ARM_NEON) || defined(__ARM_NEON__)
-bool isLatin(const std::string &str) { return isLatin_NEON(str); }
-#elif defined(__riscv) && __riscv_vector
-bool isLatin(const std::string &str) { return isLatin_RISCV(str); }
-#else
-bool isLatin(const std::string &str) { return isLatin_Baseline(str); }
-#endif
+  std::string result;
+  result.reserve(length);
+  for (size_t i = 0; i < length; ++i) {
+    result += charset[dist(rng)];
+  }
+
+  return result;
+}
 
 TEST(StringUtilTest, TestIsLatinFunctions) {
   std::string testStr = generateRandomString(100000);
-
   auto start_time = std::chrono::high_resolution_clock::now();
   bool result = isLatin(testStr);
   auto end_time = std::chrono::high_resolution_clock::now();
   auto duration = std::chrono::duration_cast<std::chrono::nanoseconds>(
                       end_time - start_time)
                       .count();
-  std::cout << "Running Time: " << duration << " ns" << std::endl;
+  FURY_LOG(INFO) << "Running Time: " << duration << " ns.";
+
   EXPECT_TRUE(result);
 }
 
 TEST(StringUtilTest, TestIsLatinLogic) {
-
-  // Test with Latin strings
+  // Test strings with only Latin characters
   EXPECT_TRUE(isLatin("Fury"));
   EXPECT_TRUE(isLatin(generateRandomString(80)));
 
-  // Test unaligned strings
-  std::string randomStr = generateRandomString(80);
-  EXPECT_TRUE(isLatin(randomStr + "1"));
-  EXPECT_TRUE(isLatin(randomStr + "12"));
-  EXPECT_TRUE(isLatin(randomStr + "123"));
+  // Test unaligned strings with only Latin characters
+  EXPECT_TRUE(isLatin(generateRandomString(80) + "1"));
+  EXPECT_TRUE(isLatin(generateRandomString(80) + "12"));
+  EXPECT_TRUE(isLatin(generateRandomString(80) + "123"));
 
-  // Test with non-Latin strings
+  // Test strings with non-Latin characters
   EXPECT_FALSE(isLatin("你好, Fury"));
-  EXPECT_FALSE(isLatin(randomStr + "你好"));
-  EXPECT_FALSE(isLatin(randomStr + "1你好"));
+  EXPECT_FALSE(isLatin(generateRandomString(80) + "你好"));
+  EXPECT_FALSE(isLatin(generateRandomString(80) + "1你好"));
   EXPECT_FALSE(isLatin(generateRandomString(11) + "你"));
   EXPECT_FALSE(isLatin(generateRandomString(10) + "你好"));
   EXPECT_FALSE(isLatin(generateRandomString(9) + "性能好"));
